Crate zng_wgt_wrap

Wrap panel, properties and nodes.

Crate

This crate is part of the zng project.

Structs

Wrap

W Wrapping inline layout.

Traits

WidgetInfoWrapExt

Extension methods for WidgetInfo that may represent a Wrap! instance.

Functions

children

c Inlined wrap items.

children_align

c Children align.

get_index

P Get the child index in the parent wrap.

get_index_len

P Get the child index and number of children.

get_rev_index

P Get the child index, starting from the last child at 0.

is_even

P If the child index is even.

is_first

P If the child is the first.

is_last

P If the child is the last.

is_odd

P If the child index is odd.

lazy_sample

Create a node that estimates the size of wrap panel children.

lazy_size

Create a node that estimates the size of wrap panel children.

node

Wrap node.

spacing

c Space in between items and rows.