Function zng_wgt_wrap::lazy_sample

pub fn lazy_sample(
    children_len: impl IntoVar<usize>,
    spacing: impl IntoVar<GridSpacing>,
    child_sample: impl UiNode,
) -> impl UiNode

Create a node that estimates the size of wrap panel children.

The estimation assumes that all items have the size of child_sample.