zng_wgt_wrap

Function spacing

pub fn spacing(wgt: &mut WidgetBuilding, spacing: impl IntoVar<GridSpacing>)

B Space in between items and rows.

Note that column space is limited for bidirectional inline items as it only inserts spacing between items once and bidirectional text can interleave items, consider using word_spacing for inline text.

Mixin Property

This property is a mixin, it modifies the widget builder during build, it does not produce a standalone UI node.