Struct zng_wgt_wrap::Wrap

pub struct Wrap(/* private fields */);

W Wrapping inline layout.

Implementations

impl Wrap

pub fn widget_new() -> Self

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Wrap

pub fn txt_align(&self, align: impl IntoVar<Align>)

P Alignment of children in this widget and of nested wrap panels and texts.

Note that this is only used for children alignment in this widget if children_align is not set on it.

pub fn line_spacing(&self, spacing: impl IntoVar<Length>)

P Space in between rows of this widget and of nested wrap panels and texts.

Note that this is only used for row spacing in this widget if spacing is not set on it.

impl Wrap

pub fn children(&self, children: impl UiNodeList)

c Inlined wrap items.

impl Wrap

pub fn spacing(&self, spacing: impl IntoVar<GridSpacing>)

c Space in between items and rows.

Note that column space is limited for bidirectional inline items as it only inserts spacing between items once and bidirectional text can interleave items, consider using word_spacing for inline text.

impl Wrap

pub fn children_align(&self, align: impl IntoVar<Align>)

c Children align.