zng_wgt_wrap

Function spacing

pub fn spacing(spacing: impl IntoVar<GridSpacing>)

c Space in between items and rows.

Note that column space is limited for bidirectional inline items as it only inserts spacing between items once and bidirectional text can interleave items, consider using word_spacing for inline text.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.